refactor(cartridge): move util functions to upper module
